add logger && change sentry(not test yet) && improve rc_controller
add logger && change sentry(not test yet) && improve rc_controller¹ 

Update README.md

补充

finish step 1&&2

delete practice

rc improve

rc_improve

add log

draw

fix bug

delete log && add socket(part

logger

fix

fix

test

日志系统历史性的一刻

fix

FIX

test

add log

change something unimportant

_

_

_

_

udp

test

_

remove cv_ctrl

looger change to std library

_

_

_

shoot_add_log

logger into_txt

draw_fric_log

完善logger but no test yet

fix logger

_

update sentry:v1

rc

!!!!!

merge many commits

DO
add logger
change sentry

---something not important--
Update README.md

补充

finish step 1&&2

delete practice

rc improve

rc_improve

add log

draw

fix bug

delete log && add socket(part

logger

fix

fix

test

日志系统历史性的一刻

fix

FIX

test

add log

change something unimportant

_

_

_

_
4

udp

test

_

remove cv_ctrl

looger change to std library

_

_

_

shoot_add_log

logger into_txt

draw_fric_log

完善logger but no test yet

fix logger

_

update sentry:v1

rc

!!!!!

+++ src/gimbal/gimbal_sentry.cc
@@ -55,7 +55,7 @@ namespace Gimbal
             } else {
                 init_stop_times = 0;
             }
-            if (init_stop_times >= Config::GIMBAL_INIT_STOP_TIME)
+            if (init_stop_times >= static_cast<int>(Config::GIMBAL_INIT_STOP_TIME))
                 robot_set->inited |= 1 << 2;
             UserLib::sleep_ms(Config::GIMBAL_CONTROL_TIME);
         }
